fix(installer): pytorch and ROCm versions are incompatible
Each version of torch is only available for specific versions of CUDA and ROCm.
The Invoke installer tries to install torch 2.4.1 with ROCm 5.6 support, which
does not exist. As a result, the installation falls back to the default CUDA
version so AMD GPUs aren't detected. This commits fixes that by bumping the
ROCm version to 6.1, as suggested by the PyTorch documentation. [1]

The specified CUDA version of 12.4 is still correct according to [1] so it does
need to be changed.

Closes #7146

[1]: https://pytorch.org/get-started/previous-versions/#v241